Python reputation API
New fix
